organizando el controlador en clases y arreglando el diagrama de clases
